% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_genlasso.R
\name{coef.quantile_genlasso}
\alias{coef.quantile_genlasso}
\title{Coef function for quantile_genlasso object}
\usage{
\method{coef}{quantile_genlasso}(obj, s = NULL)
}
\arguments{
\item{obj}{The \code{quantile_genlasso} object.}

\item{s}{Vector of integers specifying the tau and lambda values to consider
for predictions; for each \code{i} in this vector, a prediction is made at 
quantile level \code{tau[i]} and tuning parameter value \code{lambda[i]},
according to the \code{tau} and \code{lambda} vectors stored in the given  
\code{quantile_genlasso} object \code{obj}. (Said differently, \code{s} 
specifies the columns of \code{obj$beta} to use for the predictions.)
Default is NULL, which means that all tau and lambda values will be
considered.}
}
\description{
Retrieve coefficients for estimating the conditional quantiles, using the
generalized lasso coefficients at particular tau or lambda values.
}
