% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price.transformations.R
\name{getPrice}
\alias{getPrice}
\title{get price column(s) from a timeseries}
\usage{
getPrice(x, symbol = NULL, prefer = NULL, ...)
}
\arguments{
\item{x}{A data object with columns containing data to be extracted}

\item{symbol}{text string containing the symbol to extract}

\item{prefer}{preference for any particular type of price, see Details}

\item{\dots}{any other passthrough parameters}
}
\description{
Will attempt to locate price column(s) from a time series with rational defaults.
}
\details{
May be subset by symbol and preference.
\code{prefer} Preference will be for any commonly used financial time series price description,
e.g. 'trade', 'close', 'bid', 'ask' with specific tests and matching for types and column names
currently supported in R, but a default grep match will be performed if one of the supported types doesn't match.
}

