% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price.transformations.R
\name{has.Qty}
\alias{has.Ask}
\alias{has.AskSize}
\alias{has.Bid}
\alias{has.BidSize}
\alias{has.Price}
\alias{has.Qty}
\alias{has.Trade}
\alias{is.BBO}
\alias{is.TBBO}
\title{check for Trade, Bid, and Ask/Offer (BBO/TBBO), Quantity, and Price data}
\usage{
has.Qty(x, which = FALSE)
}
\arguments{
\item{x}{data object}

\item{which}{disply position of match}
}
\description{
A set of functions to check for appropriate TBBO/BBO and price column
names within a data object, as well as the availability and
position of those columns.
}

