\name{quantmod-package}
\alias{quantmod-package}
\alias{quantmod}
\alias{quantmodenv}
\docType{package}
\title{
Quantitative Financial Modelling Framework
}
\description{
Quantitative Financial Modelling and Trading Framework for R
}
\details{
\tabular{ll}{
Package: \tab quantmod\cr
Type: \tab Package\cr
Version: \tab 0.3-241\cr
Date: \tab 2012-12-31\cr
Depends: \tab xts(>=0.9-1),Defaults\cr
Suggests: \tab DBI,RMySQL,TTR,fSeries,its\cr
LazyLoad: \tab yes\cr
License: \tab GPL-3\cr
URL: \tab http://www.quantmod.com\cr
URL: \tab http://quantmod.r-forge.r-project.org\cr
}
The quantmod package for R is designed to
assist the quantitative trader in the
development, testing, and deployment
of statistically based trading models.


\emph{What quantmod IS}

A rapid prototyping environment, with comprehensive
tools for data management and visualization.
where quant traders can quickly
and cleanly explore and build trading models.


\emph{What quantmod is NOT}

A replacement for anything statistical.
It has no 'new' modelling routines or
analysis tool to speak of. It does now
offer charting not currently available
elsewhere in R, but most everything else
is more of a wrapper to what you already
know and love about the langauge
and packages you currently use.

quantmod makes modelling easier by removing
the repetitive workflow issues surrounding
data management, modelling interfaces,
and performance analysis. 

}
\author{
Jeffrey A. Ryan

Maintainer: Jeffrey A. Ryan <jeff.a.ryan@gmail.com>
}
\keyword{ package }
